requirements.txt
flask==1.0.2
psycopg2-binary
bcrypt
